+++ harbor_cli/commands/api/project.py
@@ -87,6 +87,39 @@ def get_project_info(
     render_result(p, ctx)
 
 
+# HarborAsyncClient.get_repositories()
+@app.command("repos")
+@inject_resource_options()
+def list_repos(
+    ctx: typer.Context,
+    project: Optional[str] = typer.Argument(
+        help="Name of project to fetch repositories from.",
+    ),
+    query: Optional[str] = None,
+    sort: Optional[str] = None,
+    page: int = 1,
+    page_size: int = 10,
+    limit: Optional[int] = ...,
+) -> None:
+    """List all repositories in a project.
+
+    Alternative to `repository list`"""
+    repos = state.run(
+        state.client.get_repositories(
+            project,
+            query=query,
+            sort=sort,
+            page=page,
+            page_size=page_size,
+            limit=limit,
+        ),
+        "Fetching repositories...",
+    )
+    if not repos:
+        info(f"{project!r} has no repositories.")
+    render_result(repos, ctx)
+
+
 # HarborAsyncClient.get_project_logs()
 @app.command("logs")
 @inject_resource_options()
@@ -143,7 +176,8 @@ def project_exists(
 @app.command("create", no_args_is_help=True)
 @inject_help(ProjectReq)
 @inject_help(
-    ProjectMetadata
+    ProjectMetadata,
+    remove=['The valid values are "true", "false".'],
 )  # inject this first so its "public" field takes precedence
 def create_project(
     ctx: typer.Context,
@@ -196,6 +230,11 @@ def create_project(
         None,
         "--retention-id",
     ),
+    auto_sbom_generation: Optional[bool] = typer.Option(
+        None,
+        "--sbom-generation",
+        is_flag=False,
+    ),
     # TODO: add support for adding CVE allowlist when creating a project
 ) -> None:
     """Create a new project."""
@@ -203,16 +242,20 @@ def create_project(
         project_name=project_name,
         storage_limit=storage_limit,
         registry_id=registry_id,
-        metadata=ProjectMetadata(
-            # validator does bool -> str conversion for the string bool fields
-            public=public,  # type: ignore
-            enable_content_trust=enable_content_trust,
-            enable_content_trust_cosign=enable_content_trust_cosign,
-            prevent_vul=prevent_vul,
-            severity=severity,
-            auto_scan=auto_scan,
-            reuse_sys_cve_allowlist=reuse_sys_cve_allowlist,
-            retention_id=retention_id,
+        metadata=ProjectMetadata.model_validate(
+            # NOTE: Constructing via a dict here to avoid type checking noise.
+            # See tests/api/test_models.py for more information.
+            {
+                "public": public,
+                "enable_content_trust": enable_content_trust,
+                "enable_content_trust_cosign": enable_content_trust_cosign,
+                "prevent_vul": prevent_vul,
+                "severity": severity,
+                "auto_scan": auto_scan,
+                "reuse_sys_cve_allowlist": reuse_sys_cve_allowlist,
+                "retention_id": retention_id,
+                "auto_sbom_generation": auto_sbom_generation,
+            }
         ),
     )
     location = state.run(
@@ -283,7 +326,8 @@ def list_projects(
 @app.command("update", no_args_is_help=True)
 @inject_help(ProjectReq)
 @inject_help(
-    ProjectMetadata
+    ProjectMetadata,
+    remove=['The valid values are "true", "false".'],
 )  # inject this first so its "public" field takes precedence
 def update_project(
     ctx: typer.Context,
@@ -336,6 +380,11 @@ def update_project(
         None,
         "--retention-id",
     ),
+    auto_sbom_generation: Optional[bool] = typer.Option(
+        None,
+        "--sbom-generation",
+        is_flag=False,
+    ),
 ) -> None:
     """Update project information."""
     req_params = model_params_from_ctx(ctx, ProjectReq)
@@ -346,7 +395,7 @@ def update_project(
     arg = get_project_arg(project_name_or_id)
     project = get_project(arg)
     if project.metadata is None:
-        project.metadata = ProjectMetadata()  # pyright: ignore[reportCallIssue] # mypy bug
+        project.metadata = ProjectMetadata()
 
     # Create updated models from params
     req = create_updated_model(
@@ -748,9 +797,9 @@ def list_project_members(
     entity_name: Optional[str] = typer.Option(
         None, "--entity", help="Entity name to search for."
     ),
-    page: int = ...,  # type: ignore
-    page_size: int = ...,  # type: ignore
-    limit: Optional[int] = ...,  # type: ignore
+    page: int = ...,
+    page_size: int = ...,
+    limit: Optional[int] = ...,
 ) -> None:
     """List all members of a project."""
     project_arg = get_project_arg(project_name_or_id)